feat(flowchart): enhance chatbot UI interaction and persistence logic
This commit introduces several improvements in the flowchart component.

Firstly, it adds a 'chatbotId' property to the FlowchartComponent class to hold the ID of the current Chatbot. This helps with better data management and more natural update flow in the application.

Secondly, it modifies the subscription to route parameters by storing the received chatbot ID into the new 'chatbotId'. The logic for getting the chatbot by its ID is also updated - now, after receiving the chatbot, it calls 'drawOperators()' function to reflect this change in the UI instantly.

Thirdly, it moves some method calls including 'drawOperators()' to ngAfterViewInit where it makes more sense semantically because ngAfterViewInit is typically where you put initial setup code.

Lastly, it enhances the 'submit' method by checking if the chatbot has an ID before doing a PUT request to save the chatbot. After the successful PUT request, an alert notifies the user that the chatbot has been saved successfully. This change improves user experience and ensures data consistency.