feat(chat): implement variables support in chat component
This commit adds the ability to use and manage variables within the chat module. A new variable map has been introduced to store the variables used in conversation scripts. Operators can now collect information from user responses and assign them to variables for future use.

Modifications include:
- Adding a `variables` Map to hold variable-value pairs.
- Modifying message content extraction from `messageLog` array - `.message` property is now accessed directly.
- Checking if the recently stopped operator's script includes a variable, if so, add or update the variable value in the `variables` map.
- During operator iteration, if the operator type is 'collect' and a variable is present in the script, an empty string is initially set as its value.
- Introducing a `setVariables()` method that replaces placeholders in the content with corresponding variable values if they exist.

These enhancements improve the flexibility of the chat component.

Please note that using an undeclared variable (i.e., a variable not existing in the `variables` Map) in the script will leave the placeholder as it is.