refactor(flowchart): update operator-script model and choice component
This commit removes the 'value' field from the OperatorScript interface as it is no longer needed. In the choice.component, the title of an operator now gets generated in the 'content' field of the 'script' object. Additionally, the text field of a choice option in the front-end now binds to 'option.data.script.content' instead of 'option.data.title'. These changes provide better structure and clarity by unifying title generation and usage.