feat(chat): add interactive option support in chat messages
This commit introduces interactive options into bot messages in the chat system. A new optional property `data` has been added to the Message interface which will hold any additional data related to a message.

The ChatComponent's `continueConversation` function has been enhanced to check if the current operator type is "choice". In this case, it pushes a special kind of message onto the `messageLog`. This message contains an array of options in its data property, which are the children of the current operator.

Changes have also been made in the chat.component.html file to display these options as buttons when the message type is "choice". Users can now interact with these presented options.

Note: The displayed options are not yet functional and will be linked to actions in future works.